pyproject.toml
[tool.poetry]
name = "softdesk-api"
version = "0.1.0"
description = ""
authors = ["Julien Denizot <luunz71@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
django = "5.1.2"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
pyjwt = "2.9.0"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
